instance isolation tests
