+++ packages/standalone/spy.test.js
@@ -57,3 +57,33 @@ describe('spy()', () => {
 		fm.restore();
 	});
 });
+
+
+it('error if spy() is called and no fetch defined in config', () => {
+	const fm = fetchMock.sandbox();
+	delete fm.config.fetch;
+	expect(() => fm.spy()).toThrow();
+});
+
+it("don't error if spy() is called and fetch defined in config", () => {
+	const fm = fetchMock.sandbox();
+	fm.config.fetch = originalFetch;
+	expect(() => fm.spy()).not.toThrow();
+});
+
+it('exports a properly mocked node-fetch module shape', () => {
+	// uses node-fetch default require pattern
+	const {
+		default: fetch,
+		Headers,
+		Request,
+		Response,
+	} = fetchMock.sandbox();
+
+	expect(fetch.name).toEqual('fetchMockProxy');
+	expect(new Headers()).toBeInstanceOf(fetchMock.config.Headers);
+	expect(new Request('http://a.com')).toBeInstanceOf(
+		fetchMock.config.Request,
+	);
+	expect(new Response()).toBeInstanceOf(fetchMock.config.Response);
+});
